requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
pydantic==2.9.2
pytest==8.3.3
trio
httpx~=0.27.0
boto
asyncio
orjson